Remove a js_template argument and pass a json_data argument
