--- wagtail_embed_videos/templates/wagtail_embed_videos/chooser/embed_video_chosen.js
@@ -1,4 +1,0 @@
-function(modal) {
-    modal.respond('embedVideoChosen', {{ embed_video_json|safe }});
-    modal.close();
-}
